+++ src/config.ts
@@ -1,0 +1,7 @@
+import vscode from "vscode";
+
+export const configSection = "djlint";
+
+export function getConfig(): vscode.WorkspaceConfiguration {
+  return vscode.workspace.getConfiguration(configSection);
+}