+++ src/extension.ts
@@ -1,26 +1,10 @@
-import vscode from "vscode";
-import { supportedLanguages } from "./constants";
+import type vscode from "vscode";
 import { Formatter } from "./formatter";
-import { refreshDiagnostics } from "./linter";
+import { Linter } from "./linter";
 
-export function activate(context: vscode.ExtensionContext): void {
-  const collection = vscode.languages.createDiagnosticCollection("djLint");
-  if (vscode.window.activeTextEditor) {
-    void refreshDiagnostics(
-      vscode.window.activeTextEditor.document,
-      collection
-    );
-  }
-  const diagListener = (doc: vscode.TextDocument): undefined =>
-    void refreshDiagnostics(doc, collection);
-  context.subscriptions.push(
-    vscode.workspace.onDidOpenTextDocument(diagListener),
-    vscode.workspace.onDidSaveTextDocument(diagListener),
-    vscode.workspace.onDidCloseTextDocument((doc) => collection.delete(doc.uri))
-  );
-
-  vscode.languages.registerDocumentFormattingEditProvider(
-    supportedLanguages,
-    new Formatter()
-  );
+export async function activate(
+  context: vscode.ExtensionContext
+): Promise<void> {
+  new Formatter(context).activate();
+  await new Linter(context).activate();
 }