+++ src/formatter.ts
@@ -1,9 +1,30 @@
 import vscode from "vscode";
 import { formattingArgs } from "./args";
+import { configSection, getConfig } from "./config";
 import { runDjlint } from "./runner";
-import { getConfig } from "./utils";
 
 export class Formatter implements vscode.DocumentFormattingEditProvider {
+  protected readonly context: vscode.ExtensionContext;
+  protected providerDisposable: vscode.Disposable | undefined;
+
+  constructor(context: vscode.ExtensionContext) {
+    this.context = context;
+  }
+
+  activate(): void {
+    this.register();
+
+    this.context.subscriptions.push(
+      vscode.workspace.onDidChangeConfiguration(
+        (e: vscode.ConfigurationChangeEvent) => {
+          if (e.affectsConfiguration(`${configSection}.languages`)) {
+            this.register();
+          }
+        }
+      )
+    );
+  }
+
   async provideDocumentFormattingEdits(
     document: vscode.TextDocument,
     options: vscode.FormattingOptions
@@ -12,7 +33,7 @@ export class Formatter implements vscode.DocumentFormattingEditProvider {
 
     let stdout;
     try {
-      stdout = await runDjlint(config, document, formattingArgs, options);
+      stdout = await runDjlint(document, config, formattingArgs, options);
     } catch {
       return [];
     }
@@ -24,5 +45,18 @@ export class Formatter implements vscode.DocumentFormattingEditProvider {
     const lastLineLength = document.lineAt(lastLineId).text.length;
     const range = new vscode.Range(0, 0, lastLineId, lastLineLength);
     return [vscode.TextEdit.replace(range, stdout)];
+  }
+
+  protected register(): void {
+    const languages = getConfig().get<Record<string, string>>("languages");
+    this.providerDisposable?.dispose();
+    if (languages != null) {
+      this.providerDisposable =
+        vscode.languages.registerDocumentFormattingEditProvider(
+          Object.keys(languages),
+          this
+        );
+      this.context.subscriptions.push(this.providerDisposable);
+    }
   }
 }