+++ src/utils.ts
@@ -1,5 +1,0 @@
-import vscode from "vscode";
-
-export function getConfig(): vscode.WorkspaceConfiguration {
-  return vscode.workspace.getConfiguration("djlint");
-}